Namespace adjustments
