+++ composer.json
@@ -20,9 +20,14 @@
                   "Transbank\\Onepay\\Exceptions\\": ["lib/onepay/exceptions"],
                   "Transbank\\Webpay\\" : ["lib/webpay", "lib/webpay/soap", "lib/webpay_rest", "lib/webpay_rest/webpay_plus"],
                   "Transbank\\Webpay\\WebpayPlus\\":["lib/webpay_rest/webpay_plus"],
+"Transbank\\Webpay\\WebpayPlus\\Exceptions":["lib/webpay_rest/webpay_plus/exceptions"],
                   "Transbank\\Webpay\\Exceptions\\":["lib/webpay_rest/exceptions"],
+"Transbank\\Webpay\\Oneclick":["lib/webpay_rest/oneclick", "lib/webpay_rest/oneclick/inscription", "lib/webpay_rest/oneclick/mall"],
+        "Transbank\\Webpay\\Oneclick\\Exceptions\\":["lib/webpay_rest/oneclick/exceptions", "lib/webpay_rest/oneclick/mall/exceptions"],
                   "Transbank\\Utils\\":  ["lib/utils"],
-                  "Transbank\\PatpassByWebpay\\":["lib/webpay_rest/patpass_by_webpay"]
+                  "Transbank\\Patpass\\PatpassByWebpay\\":["lib/patpass/patpass_by_webpay"],
+                  "Transbank\\Patpass\\Exceptions":["lib/patpass/exceptions"],
+                  "Transbank\\Patpass\\PatpassByWebpay\\Exceptions":["lib/patpass/patpass_by_webpay/exceptions"]
       }
     },
     "license": "BSD-3-Clause"
