+++ lib/patpass/exceptions/PatpassException.php
@@ -1,24 +1,16 @@
 <?php
 
-namespace Transbank\Webpay\Exceptions;
 
-/**
- * class TransbankException
- *
- *   basic class for Transbank related exceptions
- *
- * @package Transbank
- *
- *
- */
-class TransbankException extends \Exception
+namespace Transbank\Patpass\Exceptions;
+
+
+class PatpassException extends \Exception
 {
-
     const DEFAULT_MESSAGE = 'An error has happened, verify given parameters and try again.';
 
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
         parent::__construct($message, $code, $previous);
-    }
 
+    }
 }
