+++ lib/patpass/patpass_by_webpay/TransactionStatusResponse.php
@@ -1,50 +1,41 @@
 <?php
+namespace Transbank\Patpass\PatpassByWebpay;
 
-
-namespace Transbank\PatpassByWebpay;
-
-
-class TransactionCommitResponse
+class TransactionStatusResponse
 {
-    private $vci;
-    private $amount;
-    private $status;
-    private $buyOrder;
-    private $sessionId;
-    private $cardNumber;
-    private $accountingDate;
-    private $transactionDate;
-    private $authorizationCode;
-    private $paymentTypeCode;
-    private $responseCode;
-    private $installmentsNumber;
+    public $vci;
+    public $amount;
+    public $status;
+    public $buyOrder;
+    public $sessionId;
+    public $cardNumber;
+    public $accountingDate;
+    public $transactionDate;
+    public $authorizationCode;
+    public $paymentTypeCode;
+    public $responseCode;
+    public $installmentsAmount;
+    public $installmentsNumber;
+    public $balance;
 
-    /**
-     * TransactionCommitResponse constructor.
-     */
+
     public function __construct($json)
     {
+        $this->vci = isset($json["vci"]) ? $json["vci"] : null;
+        $this->amount = isset($json["amount"]) ? $json["amount"] : null;
+        $this->status = isset($json["status"]) ? $json["status"] : null;
+        $this->buyOrder = isset($json["buy_order"]) ? $json["buy_order"] : null;
+        $this->sessionId = isset($json["session_id"]) ? $json["session_id"] : null;
+        $this->cardNumber = isset($json["card_detail"]) ? (isset($json["card_detail"]["card_number"]) ? $json["card_detail"]["card_number"] : null) : null;
+        $this->accountingDate = isset($json["accounting_date"]) ? $json["accounting_date"] : null;
+        $this->transactionDate = isset($json["transaction_date"]) ? $json["transaction_date"] : null;
+        $this->authorizationCode = isset($json["authorization_code"]) ? $json["authorization_code"] : null;
+        $this->paymentTypeCode = isset($json["payment_type_code"]) ? $json["payment_type_code"] : null;
+        $this->responseCode = isset($json["response_code"]) ? $json["response_code"] : null;
+        $this->installmentsAmount = isset($json["installments_amount"]) ? $json["installments_amount"] : null;
 
-        $this->setVci($this->returnValueIfExists($json, 'vci'));
-        $this->setAmount($this->returnValueIfExists($json, 'amount'));
-        $this->setStatus($this->returnValueIfExists($json, 'status'));
-        $this->setBuyOrder($this->returnValueIfExists($json, 'buy_order'));
-        $this->setSessionId($this->returnValueIfExists($json, 'session_id'));
-
-        $cardDetail = $this->returnValueIfExists($json, 'card_detail');
-        $cardNumber = $this->returnValueIfExists($cardDetail, 'card_number');
-        $this->setCardNumber($cardNumber);
-        $this->setAccountingDate($this->returnValueIfExists($json, 'accounting_date'));
-        $this->setTransactionDate($this->returnValueIfExists($json, 'transaction_date'));
-        $this->setAuthorizationCode($this->returnValueIfExists($json, 'authorization_code'));
-        $this->setPaymentTypeCode($this->returnValueIfExists($json, 'payment_type_code'));
-        $this->setResponseCode($this->returnValueIfExists($json, 'response_code'));
-        $this->setInstallmentsNumber($this->returnValueIfExists($json, 'installments_number'));
-    }
-
-    public function returnValueIfExists($json, $key)
-    {
-        return isset($json[$key]) ? $json[$key] : null;
+        $this->installmentsNumber = isset($json["installments_number"]) ? $json["installments_number"] : null;
+        $this->balance = isset($json["balance"]) ? $json["balance"] : null;
     }
 
     /**
@@ -58,7 +49,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $vci
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setVci($vci)
     {
@@ -77,7 +68,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $amount
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setAmount($amount)
     {
@@ -96,7 +87,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $status
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setStatus($status)
     {
@@ -115,7 +106,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $buyOrder
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setBuyOrder($buyOrder)
     {
@@ -134,7 +125,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $sessionId
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setSessionId($sessionId)
     {
@@ -153,7 +144,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $cardNumber
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setCardNumber($cardNumber)
     {
@@ -172,7 +163,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $accountingDate
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setAccountingDate($accountingDate)
     {
@@ -191,7 +182,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $transactionDate
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setTransactionDate($transactionDate)
     {
@@ -210,7 +201,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $authorizationCode
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setAuthorizationCode($authorizationCode)
     {
@@ -229,7 +220,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $paymentTypeCode
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setPaymentTypeCode($paymentTypeCode)
     {
@@ -248,11 +239,30 @@ class TransactionCommitResponse
     /**
      * @param mixed $responseCode
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setResponseCode($responseCode)
     {
         $this->responseCode = $responseCode;
+        return $this;
+    }
+
+    /**
+     * @return mixed
+     */
+    public function getInstallmentsAmount()
+    {
+        return $this->installmentsAmount;
+    }
+
+    /**
+     * @param mixed $installmentsAmount
+     *
+     * @return TransactionStatusResponse
+     */
+    public function setInstallmentsAmount($installmentsAmount)
+    {
+        $this->installmentsAmount = $installmentsAmount;
         return $this;
     }
 
@@ -267,7 +277,7 @@ class TransactionCommitResponse
     /**
      * @param mixed $installmentsNumber
      *
-     * @return TransactionCommitResponse
+     * @return TransactionStatusResponse
      */
     public function setInstallmentsNumber($installmentsNumber)
     {
@@ -275,5 +285,22 @@ class TransactionCommitResponse
         return $this;
     }
 
+    /**
+     * @return mixed
+     */
+    public function getBalance()
+    {
+        return $this->balance;
+    }
 
+    /**
+     * @param mixed $balance
+     *
+     * @return TransactionStatusResponse
+     */
+    public function setBalance($balance)
+    {
+        $this->balance = $balance;
+        return $this;
+    }
 }
