+++ lib/patpass/patpass_by_webpay/exceptions/TransactionCreateException.php
@@ -1,11 +1,10 @@
 <?php
 
+namespace Transbank\Patpass\PatpassByWebpay\Exceptions;
 
-namespace Transbank\PatpassByWebpay\Exceptions;
+use Transbank\Patpass\Exceptions\PatpassException;
 
-use Transbank\Webpay\Exceptions\TransbankException;
-
-class TransactionCommitException extends TransbankException {
+class TransactionCreateException extends PatpassException {
     const DEFAULT_MESSAGE = 'Transaction could not be created. Please verify given parameters';
 
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
