+++ lib/webpay_rest/oneclick/exceptions/InscriptionDeleteException.php
@@ -1,10 +1,10 @@
 <?php
 
 
-namespace Transbank\Webpay\Exceptions;
+namespace Transbank\Webpay\Oneclick\Exceptions;
+use Transbank\Webpay\WebpayPlus\Exceptions\WebpayException;
 
-
-class InscriptionDeleteException extends TransbankException
+class InscriptionDeleteException extends WebpayException
 {
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
