+++ lib/webpay_rest/oneclick/inscription/InscriptionDeleteResponse.php
