+++ lib/webpay_rest/oneclick/mall/AuthorizeMallTransactionResponse.php
