+++ lib/webpay_rest/oneclick/mall/exceptions/MallRefundTransactionException.php
@@ -1,14 +1,11 @@
 <?php
+namespace Transbank\Webpay\Oneclick\Exceptions;
+use Transbank\Webpay\WebpayPlus\Exceptions\WebpayException;
 
-
-namespace Transbank\Webpay\Exceptions;
-
-
-class MallTransactionStatusException extends TransbankException
+class MallRefundTransactionException extends WebpayException
 {
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
         parent::__construct($message, $code, $previous);
     }
-
 }
