+++ lib/webpay_rest/webpay_plus/exceptions/TransactionCommitException.php
@@ -1,10 +1,10 @@
 <?php
 
 
-namespace Transbank\Webpay\Exceptions;
+namespace Transbank\Webpay\WebpayPlus\Exceptions;
 
 
-class TransactionCommitException extends TransbankException
+class TransactionCommitException extends WebpayException
 {
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
     {
