+++ lib/webpay_rest/webpay_plus/exceptions/TransactionRefundException.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Transbank\Webpay\Exceptions;
+namespace Transbank\Webpay\WebpayPlus\Exceptions;
 
 
 /**
@@ -9,7 +9,7 @@ namespace Transbank\Webpay\Exceptions;
  *
  * @package Transbank\Webpay\exceptions
  */
-class TransactionRefundException extends TransbankException {
+class TransactionRefundException extends WebpayException {
     const DEFAULT_MESSAGE = 'Transaction could not be created. Please verify given parameters';
 
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
