+++ lib/webpay_rest/webpay_plus/exceptions/TransactionStatusException.php
@@ -1,7 +1,7 @@
 <?php
 
 
-namespace Transbank\Webpay\Exceptions;
+namespace Transbank\Webpay\WebpayPlus\Exceptions;
 
 
 /**
@@ -9,7 +9,7 @@ namespace Transbank\Webpay\Exceptions;
  *
  * @package Transbank\Webpay\exceptions
  */
-class TransactionStatusException extends TransbankException {
+class TransactionStatusException extends WebpayException {
     const DEFAULT_MESSAGE = 'Transaction Status could not be confirmed. Please verify given parameters';
 
     public function __construct($message = self::DEFAULT_MESSAGE, $code = 0, $previous = null)
